Lesson6.MVP/app/requirements.txt
sqlmodel==0.0.24
bcrypt==3.2.0
passlib[bcrypt]
psycopg2-binary==2.9.10
pydantic-settings
fastapi==0.116.1 
uvicorn==0.35.0
asyncpg==0.30.0
#psycopg==3.2.9
pika==1.3.2
httpx==0.28.1
python-jose==3.5.0
streamlit==1.48.1 
requests
dotenv
python-multipart
pytest